@@ -6,76 +6,74 @@
     "start": "ng serve",
     "build": "ng build",
     "test": "ng test",
-    "lint": "ng lint embystat",
+    "lint": "ng lint",
     "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.2.14",
-    "@angular/cdk": "^8.2.3",
-    "@angular/common": "^8.2.14",
-    "@angular/compiler": "^8.2.14",
-    "@angular/core": "^8.2.14",
-    "@angular/forms": "^8.2.14",
-    "@angular/material": "^8.2.3",
-    "@angular/platform-browser": "^8.2.14",
-    "@angular/platform-browser-dynamic": "^8.2.14",
-    "@angular/router": "^8.2.14",
+    "@angular/animations": "~9.1.0",
+    "@angular/cdk": "^9.2.0",
+    "@angular/common": "~9.1.0",
+    "@angular/compiler": "~9.1.0",
+    "@angular/core": "~9.1.0",
+    "@angular/forms": "~9.1.0",
+    "@angular/localize": "^9.1.0",
+    "@angular/material": "^9.2.0",
+    "@angular/platform-browser": "~9.1.0",
+    "@angular/platform-browser-dynamic": "~9.1.0",
+    "@angular/router": "~9.1.0",
     "@aspnet/signalr": "^1.1.4",
-    "@fortawesome/angular-fontawesome": "^0.3.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.26",
-    "@fortawesome/free-solid-svg-icons": "^5.12.0",
-    "@ngrx/effects": "^8.6.0",
-    "@ngrx/store": "^8.6.0",
-    "@ngrx/store-devtools": "^8.6.0",
+    "@fortawesome/angular-fontawesome": "^0.6.1",
+    "@fortawesome/fontawesome-svg-core": "^1.2.28",
+    "@fortawesome/free-regular-svg-icons": "^5.13.0",
+    "@fortawesome/free-solid-svg-icons": "^5.13.0",
+    "@ngrx/effects": "^9.0.0",
+    "@ngrx/store": "^9.0.0",
+    "@ngrx/store-devtools": "^9.0.0",
     "@ngx-progressbar/core": "^5.3.2",
     "@ngx-progressbar/http": "^5.3.2",
-    "@ngx-translate/core": "^11.0.1",
+    "@ngx-translate/core": "^12.1.2",
     "@ngx-translate/http-loader": "^4.0.0",
     "@types/moment": "^2.13.0",
     "angular2-chartjs": "^0.5.1",
     "bootstrap": "^4.4.1",
-    "c3": "^0.5.4",
-    "core-js": "^2.6.11",
+    "c3": "^0.7.15",
+    "core-js": "^3.6.4",
     "countup.js-angular2": "^7.3.0",
-    "hammerjs": "^2.0.8",
-    "karma-sonarqube-unit-reporter": "0.0.21",
     "moment": "^2.24.0",
-    "ng-mocks": "^8.1.0",
     "ngrx-store-freeze": "^0.2.4",
     "ngx-moment": "^3.5.0",
-    "ngx-scrollbar": "^4.2.0",
-    "npm": "^6.14.3",
+    "ngx-scrollbar": "^7.2.1-beta.0",
     "rxjs": "~6.5.4",
     "rxjs-compat": "^6.5.4",
-    "ts-mockery": "^1.2.0",
+    "sonarqube-scanner": "^2.6.0",
     "tslib": "^1.10.0",
-    "uuid": "^3.3.3",
-    "wiredep": "^4.0.0",
-    "zone.js": "^0.9.1"
+    "uuid": "^7.0.2",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.24",
-    "@angular/cli": "~8.3.21",
-    "@angular/compiler-cli": "~8.2.14",
-    "@angular/language-service": "~8.2.14",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "^2.0.8",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
+    "@angular-devkit/build-angular": "~0.901.0-rc.0",
+    "@angular/cli": "~9.1.0-rc.0",
+    "@angular/compiler-cli": "~9.1.0",
+    "@angular/language-service": "~9.1.0",
+    "@types/jasmine": "~3.5.0",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^3.1.4",
-    "karma-chrome-launcher": "~2.2.0",
+    "karma": "~4.4.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~2.1.0",
+    "karma-jasmine": "~3.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
+    "karma-sonarqube-unit-reporter": "0.0.21",
     "karma-coverage": "^2.0.1",
-    "karma-coverage-istanbul-reporter": "^2.0.6",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
     "karma-junit-reporter": "^2.0.1",
-    "protractor": "~5.4.0",
-    "sonarqube-scanner": "^2.5.0",
-    "ts-node": "~7.0.0",
-    "tslint": "^5.20.1",
-    "typescript": "~3.5.3"
+    "ng-mocks": "^9.1.0",
+    "protractor": "~5.4.3",
+    "ts-node": "~8.3.0",
+    "tslint": "~6.1.0",
+    "typescript": "~3.8.3"
   }
 }
