@@ -1,5 +1,6 @@
 // Karma configuration file, see link for more information
 // https://karma-runner.github.io/1.0/config/configuration-file.html
+
 module.exports = function (config) {
   config.set({
     basePath: '',
@@ -14,9 +15,9 @@ module.exports = function (config) {
       require('karma-sonarqube-unit-reporter')
     ],
     client: {
-      clearContext: false
+      clearContext: false // leave Jasmine Spec Runner output visible in browser
     },
-    angularFilesort: {
+	angularFilesort: {
       whitelist: ['/**/!(*.html|*.spec|*.mock).js']
     },
     reporters: ['progress', 'junit', 'sonarqubeUnit'],
