@@ -3,7 +3,7 @@ import { Component, Input, OnInit } from '@angular/core';
 import { Language } from '../../models/language';
 
 @Component({
-  selector: 'app-language',
+  selector: 'es-language',
   templateUrl: './language.component.html',
   styleUrls: ['./language.component.scss']
 })
