@@ -1,9 +1,11 @@
+import { Moment } from 'moment';
+
 export class Job {
     state: number;
     currentProgressPercentage: number;
     id: string;
-    startTimeUtc: Date;
-    endTimeUtc: Date;
+    startTimeUtc: Moment;
+    endTimeUtc: Moment;
     title: string;
     description: string;
     trigger: string;
