@@ -2,7 +2,7 @@ import { Pipe, PipeTransform } from '@angular/core';
 
 @Pipe({ name: 'capitalizeFirst' })
 export class CapitalizeFirstPipe implements PipeTransform {
-  transform(value: string, args: any[]): string {
+  transform(value: string): string {
     if (!value) {
       return null;
     }
